Signed unsigned comparison.

+++ OpenSim/Common/Reporter.h
@@ -212,7 +212,8 @@ private:
             for (auto idx = 0; idx < input.getNumConnectees(); ++idx) {
                 const auto& chan = input.getChannel(idx);
                 const auto& outName = chan.getName();
-                const auto& truncName = outName.size() <= _width ?
+                const auto& truncName = 
+                    static_cast<int>(outName.size()) <= _width ?
                     outName : outName.substr(outName.size() - _width);
                 std::cout << std::setw(_width) << truncName << "|";
             }
